Move ProfilePicture to new file
