+++ CanvasPlusPlayground/Features/Profile/ProfileView.swift
@@ -108,24 +108,3 @@ struct ProfileView: View {
         }
     }
 }
-
-struct ProfilePicture: View {
-    let user: User
-
-    var body: some View {
-        AsyncImage(url: user.avatarURL) { image in
-            image
-                .resizable()
-                .clipShape(.circle)
-        } placeholder: {
-            Image(systemName: "person.circle.fill")
-                .resizable()
-                .foregroundColor(.gray)
-        }
-        .overlay {
-            Circle()
-                .stroke(lineWidth: 1)
-                .fill(.separator)
-        }
-    }
-}
